# src/main/resources/static/api-docs.yaml
openapi: 3.0.1
info:
    title: 'MilliMeeter Server'
    description: 'Just a dating app'
    license:
        name: 'Apache 2.0'
        url: 'http://springdoc.org'
    version: v0.0.1
...
servers:
    -
        url: 'http://localhost:8080'
        description: 'Generated server url'
paths:
...
    '/api/v1/messages/{id}':
        put:
            tags:
                - Message
            summary: 'Delete reaction from the message'
            operationId: deleteReactionFromMessage
            parameters:
                -
                    name: id
                    in: path
                    required: true
                    schema:
                        type: integer
                        format: int64
            responses:
                200:
                    description: OK
                    content:
                        application/hal+json:
                            examples:
                                'Deleted reaction':
                                    description: 'Deleted reaction'
                422:
                    description: 'UNPROCESSABLE ENTITY'
                    content:
                        text/plain:
                            examples:
                                'Reaction not deleted, message not exists':
                                    description: 'Reaction not deleted, message not exists'
                                    value: 'Message not exists'
                                'Reaction not deleted, profile not belongs to match':
                                    description: 'Reaction not deleted, profile not belongs to match'
                                    value: 'Profile not belongs to match'
                                'Profile not exists':
                                    description: 'Profile not exists'
                                    value: 'Profile not exists'
                        application/json:
                            examples:
                                'Not valid delete reaction request':
                                    description: 'Not valid delete reaction request'
            security:
                - bearer-key: [  ]
    '/api/v1/messages/{id}/{reaction}':
        put:
            tags:
                - Message
            summary: 'React to the message'
            operationId: reactToMessage
            parameters:
                -
                    name: id
                    in: path
                    required: true
                    schema:
                        type: integer
                        format: int64
                -
                    name: reaction
                    in: path
                    required: true
                    schema:
                        pattern: LIKE|SUPER|CARE|HAHA|WOW|CRY|WRR
                        type: string
            responses:
                200:
                    description: OK
                    content:
                        application/hal+json:
                            examples:
                                'Reacted to message':
                                    description: 'Reacted to message'
                422:
                    description: 'UNPROCESSABLE ENTITY'
                    content:
                        text/plain:
                            examples:
                                'Not reacted, message not exists':
                                    description: 'Not reacted, message not exists'
                                    value: 'Message not exists'
                                'Not reacted, profile not belongs to match':
                                    description: 'Not reacted, profile not belongs to match'
                                    value: 'Profile not belongs to match'
                                'Profile not exists':
                                    description: 'Profile not exists'
                                    value: 'Profile not exists'
                        application/json:
                            examples:
                                'Not valid react to message request':
                                    description: 'Not valid react to message request'
            security:
                -
                    bearer-key: []
...
    '/api/v1/messages/read/{matchId}':
        put:
            tags:
                - Message
            summary: 'Read messages in the conversation'
            operationId: readMessagesInConversation
            parameters:
                -
                    name: matchId
                    in: path
                    required: true
                    schema:
                        type: integer
                        format: int64
            responses:
                200:
                    description: OK
                    content:
                        text/plain:
                            examples:
                                'Messages in conversation read':
                                    description: 'Messages in conversation read'
                422:
                    description: 'UNPROCESSABLE ENTITY'
                    content:
                        text/plain:
                            examples:
                                'Not read, match not exists':
                                    description: 'Not read, match not exists'
                                    value: 'Match not exists'
                                'Not read, profile not belongs to match':
                                    description: 'Not read, profile not belongs to match'
                                    value: 'Profile not belongs to match'
                                'Profile not exists':
                                    description: 'Profile not exists'
                                    value: 'Profile not exists'
                        application/json:
                            examples:
                                'Not valid read messages in conversation request':
                                    description: 'Not valid read messages in conversation request'
            security:
                -
                    bearer-key: []
...
    /api/v1/messages:
        post:
            tags:
                - Message
            summary: 'Send the message'
            operationId: send
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/SentMessageDto'
                required: true
            responses:
                400:
                    description: 'BAD REQUEST'
                    content:
                        application/json:
                            examples:
                                'Bad request':
                                    description: 'Bad request'
                201:
                    description: CREATED
                    content:
                        application/hal+json:
                            examples:
                                'Message sent':
                                    description: 'Message sent'
                422:
                    description: 'UNPROCESSABLE ENTITY'
                    content:
                        text/plain:
                            examples:
                                'Message not sent, match not exists':
                                    description: 'Message not sent, match not exists'
                                    value: 'Match not exists'
                                'Message not sent, parent message not exists':
                                    description: 'Message not sent, parent message not exists'
                                    value: 'Parent message not exists'
                                'Message not sent, profile not belongs to match':
                                    description: 'Message not sent, profile not belongs to match'
                                    value: 'Profile not belongs to match'
                                'Profile not exists':
                                    description: 'Profile not exists'
                                    value: 'Profile not exists'
                        application/json:
                            examples:
                                'Not valid send request':
                                    description: 'Not valid send request'
            security:
                -
                    bearer-key: []
...
    /api/v1/matches:
        get:
            tags:
                - Match
            summary: 'Get matches without messages'
            operationId: findMatches
            responses:
                422:
                    description: 'UNPROCESSABLE ENTITY'
                    content:
                        text/plain:
                            examples:
                                'Profile not exists':
                                    description: 'Profile not exists'
                                    value: 'Profile not exists'
                200:
                    description: OK
                    content:
                        application/hal+json:
                            examples:
                                'No matches found':
                                    description: 'No matches found'
                                'Matches found':
                                    description: 'Matches found'
                                    value: FIND_MATCHES_OK_WITH_MATCHES_RESPONSE
            security:
                -
                    bearer-key: []
    /api/v1/conversations:
        get:
            tags:
                - Match
            summary: 'Get matches with messages'
            operationId: findMatchesWithMessages
            responses:
                422:
                    description: 'UNPROCESSABLE ENTITY'
                    content:
                        text/plain:
                            examples:
                                'Profile not exists':
                                    description: 'Profile not exists'
                                    value: 'Profile not exists'
                200:
                    description: OK
                    content:
                        application/hal+json:
                            examples:
                                'No matches with messages found':
                                    description: 'No matches with messages found'
                                'Found matches with messages':
                                    description: 'Found matches with messages'
            security:
                -
                    bearer-key: []
    '/api/v1/conversations/{id}':
        get:
            tags:
                - Message
            summary: 'Get the conversation by match id'
            operationId: findMessagesByMatchId
            parameters:
                -
                    name: id
                    in: path
                    required: true
                    schema:
                        type: integer
                        format: int64
            responses:
                200:
                    description: OK
                    content:
                        application/hal+json:
                            examples:
                                'Got conversation with messages':
                                    description: 'Got conversation with messages'
                                'Got conversation without messages':
                                    description: 'Got conversation without messages'
                422:
                    description: 'UNPROCESSABLE ENTITY'
                    content:
                        text/plain:
                            examples:
                                'Messages not found, match not exists':
                                    description: 'Messages not found, match not exists'
                                    value: 'Match not exists'
                                'Messages not found, profile not belongs to match':
                                    description: 'Messages not found, profile not belongs to match'
                                    value: 'Profile not belongs to match'
                                'Profile not exists':
                                    description: 'Profile not exists'
                                    value: 'Profile not exists'
                        application/json:
                            examples:
                                'Not valid find conversation messages request':
                                    description: 'Not valid find conversation messages request'
            security:
                -
                    bearer-key: []
    '/api/v1/matches/{id}':
        delete:
            tags:
                - Match
            summary: 'Delete match by the id'
            operationId: deleteMatch
            parameters:
                -
                    name: id
                    in: path
                    required: true
                    schema:
                        type: integer
                        format: int64
            responses:
                204:
                    description: 'NO CONTENT'
                    content:
                        text/plain:
                            examples:
                                'Match deleted':
                                    description: 'Match deleted'
                422:
                    description: 'UNPROCESSABLE ENTITY'
                    content:
                        text/plain:
                            examples:
                                'Match not exists':
                                    description: 'Match not exists'
                                    value: 'Match not exists'
                                'Profile not exists':
                                    description: 'Profile not exists'
                                    value: 'Profile not exists'
                        application/json:
                            examples:
                                'Not valid delete request':
                                    description: 'Not valid delete request'
            security:
                -
                    bearer-key: []
components:
    schemas:
...
        SentMessageDto:
            required:
                - content
                - matchId
            type: object
            properties:
                matchId:
                    type: integer
                    format: int64
                content:
                    maxLength: 1000
                    minLength: 1
                    type: string
                parentMessageId:
                    type: integer
                    format: int64
...
    securitySchemes:
        bearer-key:
            type: http
            scheme: bearer
            bearerFormat: JWT